// clients/js-kit/src/plugins/types.ts
export type BasePlugin = {
  authority: PluginAuthority;
  offset?: bigint;
};

// Note: RoyaltiesPlugin is defined in royalties.ts with a more sophisticated type
export type FreezeDelegatePlugin = BasePlugin & FreezeDelegate;
export type BurnDelegatePlugin = BasePlugin & BurnDelegate;
export type TransferDelegatePlugin = BasePlugin & TransferDelegate;
export type UpdateDelegatePlugin = BasePlugin & UpdateDelegate;
export type PermanentFreezeDelegatePlugin = BasePlugin & PermanentFreezeDelegate;
export type AttributesPlugin = BasePlugin & Attributes;
export type PermanentTransferDelegatePlugin = BasePlugin & PermanentTransferDelegate;
export type PermanentBurnDelegatePlugin = BasePlugin & PermanentBurnDelegate;
export type EditionPlugin = BasePlugin & Edition;
export type MasterEditionPlugin = BasePlugin & MasterEdition;
export type AddBlockerPlugin = BasePlugin & AddBlocker;
export type ImmutableMetadataPlugin = BasePlugin & ImmutableMetadata;
export type VerifiedCreatorsPlugin = BasePlugin & VerifiedCreators;
export type AutographPlugin = BasePlugin & Autograph;
export type BubblegumV2Plugin = BasePlugin & BubblegumV2;
export type FreezeExecutePlugin = BasePlugin & FreezeExecute;
export type PermanentFreezeExecutePlugin = BasePlugin & PermanentFreezeExecute;

export type CommonPluginsList = {
  attributes?: AttributesPlugin;
  royalties?: RoyaltiesPlugin;
  updateDelegate?: UpdateDelegatePlugin;
  permanentFreezeDelegate?: PermanentFreezeDelegatePlugin;
  permanentTransferDelegate?: PermanentTransferDelegatePlugin;
  permanentBurnDelegate?: PermanentBurnDelegatePlugin;
  addBlocker?: AddBlockerPlugin;
  immutableMetadata?: ImmutableMetadataPlugin;
  autograph?: AutographPlugin;
  verifiedCreators?: VerifiedCreatorsPlugin;
  freezeExecute?: FreezeExecutePlugin;
  permanentFreezeExecute?: PermanentFreezeExecutePlugin;
};

export type AssetPluginsList = {
  freezeDelegate?: FreezeDelegatePlugin;
  burnDelegate?: BurnDelegatePlugin;
  transferDelegate?: TransferDelegatePlugin;
  edition?: EditionPlugin;
} & CommonPluginsList;

export type CollectionPluginsList = {
  masterEdition?: MasterEditionPlugin;
  bubblegumV2?: BubblegumV2Plugin;
} & CommonPluginsList;